Updated app to include and observe
App updated to include observes and prevent re-rendering of map when zoomed into location once select feature for percentage is triggered.

In addition map() function now applied at the start of app to read in all data sheets. However the sheets are not yet utilised and this is in preparation for future development and a method for using map from the purrr package.