Add files list to update LA output with diff data
Added an update to select input which gives different options for outputs at LA level depending on the data. The fields available for plotting updates depending on the data in the list selected from the top select input.

There is some code for adding a second observe to render the plot after percentages option is un-ticked but this currently does not work and is to be fixed in the future.